# [BesTLA] First-token inference optimization