--- a/bestla/bestla/bestla_epilogue.h
+++ b/bestla/bestla/bestla_epilogue.h
@@ -23,168 +23,6 @@ namespace bestla {
 namespace epilogue {
 namespace gemm {
 
-template <typename DT>
-struct ParamAccumulatorWriteBack {
-  DT* C;
-  int ldc;
-  void* elt_const_v;
-};
-
-template <BTLA_ISA ISA_T, typename _SRC_T, typename _DST_T>
-class AccumulatorWriteBack {
- public:
-  using SType = _SRC_T;
-  using DType = _DST_T;
-  using Param = ParamAccumulatorWriteBack<DType>;
-
-  static BTLA_CODE forward(const _SRC_T* cacheptr, const int cachestep, const int M_offset, const int N_offset,
-                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
-    auto COffset = M_offset * _param.ldc + N_offset;
-    auto cptr = _param.C + COffset;
-    if constexpr (std::is_same_v<_SRC_T, DType>) {
-      if (cacheptr == cptr) {
-        return BTLA_CODE::Success;
-      }
-    }
-
-    return kernel::wrapper::Memcpy2D::template forward<ISA_T, SType, DType>(cacheptr, cptr, M, N, cachestep, _param.ldc,
-                                                                            _param.elt_const_v);
-  }
-};
-
-template <BTLA_ISA ISA_T, typename _SRC_T, typename _DST_T, BTLA_ELTWISEOP _OP>
-class CustomAccumulatorWriteBackWithEltop {
- public:
-  using Param = ParamAccumulatorWriteBack<_DST_T>;
-  static BTLA_CODE forward(const _SRC_T* cacheptr, const int cachestep, const int M_offset, const int N_offset,
-                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
-    auto COffset = M_offset * _param.ldc + N_offset;
-    auto cptr = _param.C + COffset;
-    if constexpr (std::is_same<_SRC_T, float>::value && std::is_same<_DST_T, float>::value) {
-      return kernel::wrapper::Memcpy2D::template forward1<ISA_T, float, float, _OP>(cacheptr, cptr, M, N, cachestep,
-                                                                                    _param.ldc, _param.elt_const_v);
-    } else {
-      assert(false);
-    }
-  }
-};
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackFp32 = AccumulatorWriteBack<ISA_T, float, float>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackInt32 = AccumulatorWriteBack<ISA_T, int, int>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackBf16 = AccumulatorWriteBack<ISA_T, utils::bf16, utils::bf16>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackFp16 = AccumulatorWriteBack<ISA_T, utils::fp16, utils::fp16>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackBf16Fp32 = AccumulatorWriteBack<ISA_T, utils::bf16, float>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackFp16Fp32 = AccumulatorWriteBack<ISA_T, utils::fp16, float>;
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackFp32Bf16 = AccumulatorWriteBack<ISA_T, float, utils::bf16>;
-
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackWithGeluFp32 = CustomAccumulatorWriteBackWithEltop<ISA_T, float, float, BTLA_ELTWISEOP::GELU>;
-
-template <BTLA_ISA ISA_T>
-using AccumulatorWriteBackWithSwishFp32 =
-    CustomAccumulatorWriteBackWithEltop<ISA_T, float, float, BTLA_ELTWISEOP::SWISH>;
-
-template <typename DT>
-struct ParamAlphaBetaProcess {
-  DT *C, *D;
-  int ldc, ldd;
-  float alpha, beta;
-};
-template <BTLA_ISA ISA_T>
-class AlphaBetaProcessFp32 {
- public:
-  using Param = ParamAlphaBetaProcess<float>;
-
-  static BTLA_CODE forward(const float* cacheptr, const int cachestep, const int M_offset, const int N_offset,
-                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
-    auto DOffset = M_offset * _param.ldd + N_offset;
-    auto COffset = M_offset * _param.ldc + N_offset;
-    auto cptr = _param.C + COffset;
-    auto dptr = _param.D + DOffset;
-    return kernel::wrapper::AlphaBetaF32F32::template forward<ISA_T>(_param.alpha, cacheptr, cachestep, _param.beta,
-                                                                     dptr, _param.ldd, cptr, _param.ldc, M, N);
-  }
-};
-
-struct ParamCompFp32BlockEpilogue {
-  void* scales;
-  BTLA_DTYPE scaledtype;
-  int ldsb;
-  int8_t* zps = nullptr;
-  float* reduce = nullptr;
-  int ldra;
-};
-template <BTLA_ISA ISA_T>
-class CompFp32BlockEpilogue {
- public:
-  using Param = ParamCompFp32BlockEpilogue;
-  static BTLA_CODE forward(const float* srcptr, float* dstptr, const int cachestep, const int M_offset,
-                           const int N_offset, const int K_offset, const int M, const int N, const Param& _param,
-                           void* tmpcache, size_t cachesize) {
-    auto ret = BTLA_CODE::NotSupport;
-    if (_param.scaledtype == BTLA_DTYPE::F32) {
-      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
-          reinterpret_cast<float*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
-          cachestep, M, N);
-      assert(ret == BTLA_CODE::Success);
-      if (_param.zps != nullptr) {
-        ret = kernel::wrapper::RemoveZeroPointBias::forward_wei<ISA_T>(
-            dstptr, cachestep, M, N, _param.zps + K_offset * _param.ldsb + N_offset,
-            reinterpret_cast<float*>(_param.scales) + K_offset * _param.ldsb + N_offset, _param.ldra,
-            _param.reduce + M_offset * _param.ldra + K_offset);
-      }
-      assert(ret == BTLA_CODE::Success);
-      return ret;
-    } else if (_param.scaledtype == BTLA_DTYPE::BF16) {
-      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
-          reinterpret_cast<utils::bf16*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
-          cachestep, M, N);
-      if (_param.zps != nullptr) {
-        assert(0);
-      }
-      assert(ret == BTLA_CODE::Success);
-      return ret;
-    } else if (_param.scaledtype == BTLA_DTYPE::F8_E8M0) {
-      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
-          reinterpret_cast<utils::f8*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
-          cachestep, M, N);
-      if (_param.zps != nullptr) {
-        assert(0);
-      }
-    } else {
-      assert(0);
-    }
-    return BTLA_CODE::NotSupport;
-  }
-};
-
-struct ParamDequantInt32ToFp32 {
-  float* C;
-  int ldc;
-  int ldsa;
-  float* scalesA;
-  float* scalesB;
-};
-template <BTLA_ISA ISA_T>
-class DequantInt32ToFp32 {
- public:
-  using Param = ParamDequantInt32ToFp32;
-  static BTLA_CODE forward(const int32_t* cacheptr, const int cachestep, const int M_offset, const int N_offset,
-                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
-    auto COffset = M_offset * _param.ldc + N_offset;
-    auto cptr = _param.C + COffset;
-    return kernel::wrapper::DequanS32Fp32::template forward<ISA_T>(cacheptr, cachestep, cptr, _param.ldc, M, N,
-                                                                   _param.scalesA + M_offset * _param.ldsa, _param.ldsa,
-                                                                   _param.scalesB + N_offset);
-  }
-};
-
 struct ParamPcKBlockCompInt8Epilogue {
   void* scalesB;
   BTLA_DTYPE scaleBdtype;
@@ -198,14 +36,16 @@ struct ParamPcKBlockCompInt8Epilogue {
   float* reduceA = nullptr;
   int K = 1;
 };
-template <class Fp32Epilogue, BTLA_ISA ISA_T>
+template <class Fp32Epilogue>
 class PcKBlockCompInt8Epilogue {
  public:
+  using Fp32Param = typename Fp32Epilogue::Param;
   struct Param {
     ParamPcKBlockCompInt8Epilogue param1;
-    typename Fp32Epilogue::Param param2;
+    Fp32Param param2;
   };
   using Fp32Epi = Fp32Epilogue;
+  template <BTLA_ISA ISA_T>
   static BTLA_CODE forward(const int32_t* srcptr, const int cachestep, const int M_offset, const int N_offset,
                            const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
     BTLA_CODE ret = BTLA_CODE::NotSupport;
@@ -256,9 +96,166 @@ class PcKBlockCompInt8Epilogue {
             1, param1.K, param1.reduceA + M_offset, redb);
       }
     }
-    Fp32Epilogue::forward(tmpfp32ptr, cachestep, M_offset, N_offset, M, N, _param.param2, tmpcache, cachesize);
+    Fp32Epilogue::forward<ISA_T>(tmpfp32ptr, cachestep, M_offset, N_offset, M, N, _param.param2, tmpcache, cachesize);
 
     return ret;
+  }
+};
+
+template <typename DT>
+struct ParamAccumulatorWriteBack {
+  DT* C;
+  int ldc;
+  void* elt_const_v;
+};
+
+template <typename _SRC_T, typename _DST_T>
+class AccumulatorWriteBack {
+ public:
+  using SType = _SRC_T;
+  using DType = _DST_T;
+  using Param = ParamAccumulatorWriteBack<DType>;
+  using PcCompInt8Epi = bestla::epilogue::gemm::PcKBlockCompInt8Epilogue<AccumulatorWriteBack<_SRC_T, _DST_T>>;
+
+  template <BTLA_ISA ISA_T>
+  static BTLA_CODE forward(const _SRC_T* cacheptr, const int cachestep, const int M_offset, const int N_offset,
+                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
+    auto COffset = M_offset * _param.ldc + N_offset;
+    auto cptr = _param.C + COffset;
+    if constexpr (std::is_same_v<_SRC_T, DType>) {
+      if (cacheptr == cptr) {
+        return BTLA_CODE::Success;
+      }
+    }
+
+    return kernel::wrapper::Memcpy2D::template forward<ISA_T, SType, DType>(cacheptr, cptr, M, N, cachestep, _param.ldc,
+                                                                            _param.elt_const_v);
+  }
+};
+
+template <typename _SRC_T, typename _DST_T, BTLA_ELTWISEOP _OP>
+class CustomAccumulatorWriteBackWithEltop {
+ public:
+  using PcCompInt8Epi =
+      bestla::epilogue::gemm::PcKBlockCompInt8Epilogue<CustomAccumulatorWriteBackWithEltop<_SRC_T, _DST_T, _OP>>;
+  using Param = ParamAccumulatorWriteBack<_DST_T>;
+  template <BTLA_ISA ISA_T>
+  static BTLA_CODE forward(const _SRC_T* cacheptr, const int cachestep, const int M_offset, const int N_offset,
+                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
+    auto COffset = M_offset * _param.ldc + N_offset;
+    auto cptr = _param.C + COffset;
+    if constexpr (std::is_same<_SRC_T, float>::value && std::is_same<_DST_T, float>::value) {
+      return kernel::wrapper::Memcpy2D::template forward1<ISA_T, float, float, _OP>(cacheptr, cptr, M, N, cachestep,
+                                                                                    _param.ldc, _param.elt_const_v);
+    } else {
+      assert(false);
+    }
+  }
+};
+using AccumulatorWriteBackFp32 = AccumulatorWriteBack<float, float>;
+using AccumulatorWriteBackInt32 = AccumulatorWriteBack<int, int>;
+using AccumulatorWriteBackBf16 = AccumulatorWriteBack<utils::bf16, utils::bf16>;
+using AccumulatorWriteBackFp16 = AccumulatorWriteBack<utils::fp16, utils::fp16>;
+using AccumulatorWriteBackBf16Fp32 = AccumulatorWriteBack<utils::bf16, float>;
+using AccumulatorWriteBackFp16Fp32 = AccumulatorWriteBack<utils::fp16, float>;
+using AccumulatorWriteBackFp32Bf16 = AccumulatorWriteBack<float, utils::bf16>;
+
+using AccumulatorWriteBackWithGeluFp32 = CustomAccumulatorWriteBackWithEltop<float, float, BTLA_ELTWISEOP::GELU>;
+
+using AccumulatorWriteBackWithSwishFp32 = CustomAccumulatorWriteBackWithEltop<float, float, BTLA_ELTWISEOP::SWISH>;
+
+template <typename DT>
+struct ParamAlphaBetaProcess {
+  DT *C, *D;
+  int ldc, ldd;
+  float alpha, beta;
+};
+class AlphaBetaProcessFp32 {
+ public:
+  using Param = ParamAlphaBetaProcess<float>;
+
+  template <BTLA_ISA ISA_T>
+  static BTLA_CODE forward(const float* cacheptr, const int cachestep, const int M_offset, const int N_offset,
+                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
+    auto DOffset = M_offset * _param.ldd + N_offset;
+    auto COffset = M_offset * _param.ldc + N_offset;
+    auto cptr = _param.C + COffset;
+    auto dptr = _param.D + DOffset;
+    return kernel::wrapper::AlphaBetaF32F32::template forward<ISA_T>(_param.alpha, cacheptr, cachestep, _param.beta,
+                                                                     dptr, _param.ldd, cptr, _param.ldc, M, N);
+  }
+};
+
+struct ParamCompFp32BlockEpilogue {
+  void* scales;
+  BTLA_DTYPE scaledtype;
+  int ldsb;
+  int8_t* zps = nullptr;
+  float* reduce = nullptr;
+  int ldra;
+};
+class CompFp32BlockEpilogue {
+ public:
+  using Param = ParamCompFp32BlockEpilogue;
+  template <BTLA_ISA ISA_T>
+  static BTLA_CODE forward(const float* srcptr, float* dstptr, const int cachestep, const int M_offset,
+                           const int N_offset, const int K_offset, const int M, const int N, const Param& _param,
+                           void* tmpcache, size_t cachesize) {
+    auto ret = BTLA_CODE::NotSupport;
+    if (_param.scaledtype == BTLA_DTYPE::F32) {
+      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
+          reinterpret_cast<float*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
+          cachestep, M, N);
+      assert(ret == BTLA_CODE::Success);
+      if (_param.zps != nullptr) {
+        ret = kernel::wrapper::RemoveZeroPointBias::forward_wei<ISA_T>(
+            dstptr, cachestep, M, N, _param.zps + K_offset * _param.ldsb + N_offset,
+            reinterpret_cast<float*>(_param.scales) + K_offset * _param.ldsb + N_offset, _param.ldra,
+            _param.reduce + M_offset * _param.ldra + K_offset);
+      }
+      assert(ret == BTLA_CODE::Success);
+      return ret;
+    } else if (_param.scaledtype == BTLA_DTYPE::BF16) {
+      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
+          reinterpret_cast<utils::bf16*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
+          cachestep, M, N);
+      if (_param.zps != nullptr) {
+        assert(0);
+      }
+      assert(ret == BTLA_CODE::Success);
+      return ret;
+    } else if (_param.scaledtype == BTLA_DTYPE::F8_E8M0) {
+      ret = kernel::wrapper::CompFp32BlockScale::template forward<ISA_T>(
+          reinterpret_cast<utils::f8*>(_param.scales) + K_offset * _param.ldsb + N_offset, srcptr, cachestep, dstptr,
+          cachestep, M, N);
+      if (_param.zps != nullptr) {
+        assert(0);
+      }
+    } else {
+      assert(0);
+    }
+    return BTLA_CODE::NotSupport;
+  }
+};
+
+struct ParamDequantInt32ToFp32 {
+  float* C;
+  int ldc;
+  int ldsa;
+  float* scalesA;
+  float* scalesB;
+};
+class DequantInt32ToFp32 {
+ public:
+  using Param = ParamDequantInt32ToFp32;
+  template <BTLA_ISA ISA_T>
+  static BTLA_CODE forward(const int32_t* cacheptr, const int cachestep, const int M_offset, const int N_offset,
+                           const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
+    auto COffset = M_offset * _param.ldc + N_offset;
+    auto cptr = _param.C + COffset;
+    return kernel::wrapper::DequanS32Fp32::template forward<ISA_T>(cacheptr, cachestep, cptr, _param.ldc, M, N,
+                                                                   _param.scalesA + M_offset * _param.ldsa, _param.ldsa,
+                                                                   _param.scalesB + N_offset);
   }
 };
 
@@ -277,10 +274,10 @@ struct ParamZpDequantInt32ToFp32 {
   float* reduceA = nullptr;
   int K = 1;
 };
-template <BTLA_ISA ISA_T>
 class ZpDequantInt32ToFp32 {
  public:
   using Param = ParamZpDequantInt32ToFp32;
+  template <BTLA_ISA ISA_T>
   static BTLA_CODE forward(const int32_t* cacheptr, const int cachestep, const int M_offset, const int N_offset,
                            const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
     auto COffset = M_offset * _param.ldc + N_offset;
@@ -318,10 +315,10 @@ struct ParamAlphaBetaProcessS32U8 {
   float scaleAcc, scaleC;
   int zpC;
 };
-template <BTLA_ISA ISA_T>
 class AlphaBetaProcessS32U8 {
  public:
   using Param = ParamAlphaBetaProcessS32U8;
+  template <BTLA_ISA ISA_T>
   static BTLA_CODE forward(const int32_t* cacheptr, const int cachestep, const int M_offset, const int N_offset,
                            const int M, const int N, const Param& _param, void* tmpcache, size_t cachesize) {
     auto COffset = M_offset * _param.ldc + N_offset;