@yuchengliu1 2024-05-30
There also `add` and `mul` in `custom::epilogue`.  Should we keep one `add` and `mul`?