@luoyu-intel 2024-05-30
> There also `add` and `mul` in `custom::epilogue`. Should we keep one `add` and `mul`?

`custom::epilogue` are using the ref code, they should call kernel::wrapper::Add and Mul