@luoyu-intel 2024-05-31
Windows server can't connect to the proxy server. Windows build is verified on a local windows machine.